app/requirements.txt
Flask
Flask-Caching
google-cloud-bigquery